# swagger.yaml
openapi: 3.0.0
info:
  title: ZOZO API
  contact: {}
  version: '1.0'
servers:
- url: https://busy-red-zebra-robe.cyclic.app/api/v1
  variables: {}
paths:
...
  /orders/create-payment-intent:
    post:
      tags:
      - Orders
      summary: Create Payment Intent
      operationId: CreatePaymentIntent
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/CreatePaymentIntentRequest'
              - example:
                  orderId: '123'
            example:
              orderId: '123'
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
  /orders:
    get:
      tags:
      - Orders
      summary: show my order
      operationId: showmyorder
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    post:
      tags:
      - Orders
      summary: create order
      operationId: createorder
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/createorderrequest'
              - example:
                  status: cart
                  subtotal: 120
                  shippingFee: 10
                  total: 130
                  items:
                  - title: теперь 1 фотка
                    price: 120
                    amount: 2
                    product: 63a713524a65ca6f4dce03ea
                    image: /uploads/products/7.webp
            example:
              status: cart
              subtotal: 120
              shippingFee: 10
              total: 130
              items:
              - title: теперь 1 фотка
                price: 120
                amount: 2
                product: 63a713524a65ca6f4dce03ea
                image: /uploads/products/7.webp
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
    delete:
      tags:
      - Orders
      summary: delete order
      operationId: deleteorder
      parameters: []
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
  /orders/{orderID}:
    parameters:
      - in: path
        name: orderID
        schema:
          type: string
        required: true
        description: The order ID
    patch:
      tags:
      - Orders
      summary: update order
      operationId: updateorder
      parameters: []
      requestBody:
        description: ''
        content:
          application/json:
            schema:
              allOf:
              - $ref: '#/components/schemas/updateorderrequest'
              - {}
        required: true
      responses:
        '200':
          description: ''
          headers: {}
      deprecated: false
...
components:
  schemas:
...
    CreatePaymentIntentRequest:
      title: CreatePaymentIntentRequest
      required:
      - orderId
      type: object
      properties:
        orderId:
          type: string
      example:
        orderId: '123'
...
    createorderrequest:
      title: createorderrequest
      required:
      - status
      - subtotal
      - shippingFee
      - total
      - items
      type: object
      properties:
        status:
          type: string
        subtotal:
          type: integer
          format: int32
        shippingFee:
          type: integer
          format: int32
        total:
          type: integer
          format: int32
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
          description: ''
      example:
        status: cart
        subtotal: 120
        shippingFee: 10
        total: 130
        items:
        - title: теперь 1 фотка
          price: 120
          amount: 2
          product: 63a713524a65ca6f4dce03ea
          image: /uploads/products/7.webp
    Item:
      title: Item
      required:
      - title
      - price
      - amount
      - product
      - image
      type: object
      properties:
        title:
          type: string
        price:
          type: integer
          format: int32
        amount:
          type: integer
          format: int32
        product:
          type: string
        image:
          type: string
      example:
        title: теперь 1 фотка
        price: 120
        amount: 2
        product: 63a713524a65ca6f4dce03ea
        image: /uploads/products/7.webp
    updateorderrequest:
      title: updateorderrequest
      required:
      - status
      - subtotal
      - shippingFee
      - total
      - items
      - discounts
      type: object
      properties:
        status:
          type: string
        subtotal:
          type: number
        shippingFee:
          type: number
        total:
          type: number
        items:
          type: array
          items:
            $ref: '#/components/schemas/Item'
          description: ''
        discounts:
          type: array
          items:
            $ref: '#/components/schemas/Discount'
          description: ''
    Discount:
      title: Discount
      required:
      - value
      - type
      type: object
      properties:
        value:
          oneOf:
          - type: number
          - type: integer
            format: int32
        type:
          type: string
      example:
        value: 0.99
        type: percentage
...
security:
- {}
...